Updating the US Senate's membership.
The growth in population and shifts in power have rendered the US Senate ineffective in the preservation of US Democracy and is failing to sustain the ideals upon which our country was founded. This change addresses that clear and present danger to the future of the United States of America.